vendor: Import r3137 from http://svn.netlabs.org/repos/kbuild/trunk.
Directories removed when importing:
- /dist
- /kBuild/bin
- /SlickEdit

Note that this also brings /src/lib/kStuff which is an external
repository of its own (http://trac.netlabs.org/kstuff/) but we will
keep it here instead of making a git submodule as there is no
kstuff git repo yet.

+++ src/kmk/misc.c
@@ -368,7 +368,7 @@ error (flocp, fmt, va_alist)
   VA_START (args, fmt);
   cchMsg += cchUser = vsnprintf (&szMsg[cchMsg], sizeof(szMsg) - cchMsg, fmt, args);
   VA_END (args);
-  if (   cchMsg < sizeof(szMsg)
+  if (   cchMsg < (int)sizeof(szMsg)
       && cchUser >= 0)
     {
       extern size_t maybe_con_fwrite(void const *, size_t, size_t, FILE *);
@@ -433,7 +433,7 @@ fatal (flocp, fmt, va_alist)
   VA_START (args, fmt);
   cchMsg += cchUser = vsnprintf (&szMsg[cchMsg], sizeof(szMsg) - cchMsg, fmt, args);
   VA_END (args);
-  if (   cchMsg + cchStop <= sizeof(szMsg)
+  if (   cchMsg + cchStop <= (int)sizeof(szMsg)
       && cchUser >= 0)
     {
       extern size_t maybe_con_fwrite(void const *, size_t, size_t, FILE *);
